Pre-requisites to allow later move to docker buildx
We were grabbing `jx-release-version` from a `docker` image which unfortunately is not multi-arch.
Following advice from Hervé Le Meur, I replaced this `FROM` with a direct installation from the
`github` release process to have `jx-release-version` available for this image.
I gave a try at `docker buildx build` with several platforms. `ubuntu 22.04` `docker` image is
available for :
 * linux/amd64
 * linux/arm/v7
 * linux/arm64/v8
 * linux/ppc64le
 * linux/riscv64
 * linux/s390x

 Unfortunately, `jenkins/inbound-agent:4.13-2-jdk11` is only available for:
  * linux/amd64
  * linux/arm64
  * linux/s390x

Furthermore, `gh` is not available for `s390x`, which lets only `amd64` and `arm64` for this image.
¯\_(ツ)_/¯