chore: retrieve jx-release-version from github instead of its docker image (#41)
The goal of this PR is to allow later on to build an `aarch64` `docker` image so that we can build on Oracle `aarch64` free tier,  Amazon Graviton 3, and so on...
The first step (for me, correct me if I'm wrong) is to allow the build for several architectures thanks to `buildx`.
Before changing the `Makefile`, I think we should get rid of `docker` images that are not multi-arch.

We were grabbing `jx-release-version` from a `docker` image which unfortunately is not yet multi-arch.
Following advice from @lemeurherve, I replaced this `FROM` with a direct installation from the `github` release process to have `jx-release-version` available for this image (the tool is released for several architectures (`amd64`, `arm64`, `arm32`).

I gave a try at `docker buildx build` with several platforms. 

`ubuntu 22.04` `docker` image is available for :
 * `linux/amd64`
 * `linux/arm/v7`
 * `linux/arm64/v8`
 * `linux/ppc64le`
 * `linux/riscv64`
 * `linux/s390x`

 Unfortunately, `jenkins/inbound-agent:4.13-2-jdk11` is only available for:
  * `linux/amd64`
  * `linux/arm64`
  * `linux/s390x`

`jx-release-version` is only available for :
 * `amd64`
 * `arm64`
 * `arm32`

Furthermore, `gh` is not available for `s390x`, which lets only `amd64` and `arm64` for this image.

While I was there, I changed the dependency of `jx-release-version` to `2.5.2`.